Project/Spotify/src/bg/sofia/uni/fmi/mjt/spotify/user/PasswordHash.java
package bg.sofia.uni.fmi.mjt.spotify.user;

import bg.sofia.uni.fmi.mjt.spotify.logger.ExceptionLogger;

import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;
import java.util.Base64;

public class PasswordHash {
    public static String hashPassword(String password) {
        try {
            MessageDigest md = MessageDigest.getInstance("SHA-256");

            md.update(password.getBytes());

            byte[] hashedBytes = md.digest();

            return Base64.getEncoder().encodeToString(hashedBytes);
        } catch (NoSuchAlgorithmException exception) {
            ExceptionLogger.logException(exception);
            throw new RuntimeException("Problem occurred while hashing the password.", exception);
        }
    }
}
